bug fix in post_ensight (introduced more than 10000 revision ago): we wrote node ids as float although int is expected by the EnSight Gold Format --> surprisingly, neither Paraview nor ens_checker did complain; other visualization tools do --> this bug fix does not change the output in Paraview
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@18955 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/post_ensight/post_drt_ensight_writer.cpp
@@ -2011,7 +2011,7 @@ void EnsightWriter::WriteCoordinatesForPolynomialShapefunctions
       // first write node global ids (default)
       for (int inode=0; inode<proc0map->NumGlobalElements(); ++inode)
       {
-        Write(geofile,static_cast<float>(proc0map->GID(inode))+1);
+        Write(geofile,proc0map->GID(inode)+1);
         // gid+1 delivers the node numbering of the *.dat file starting with 1
       }
     }
